requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.2
beautifulsoup4==4.12.3
requests==2.32.3
langchain==0.2.1
langchain-community==0.2.1
chromadb==0.5.0
groq==0.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.7.2
pydantic-settings==2.2.1
